[jollacomponents-internal] Use dialogs instead of plain pages when creating an account. Contributes to JB#4853
Remove AccountCreationPage as the account creation flow should use a
dialog flow. Additionally the dialog flow uses
Dialog::acceptDestination to move between pages.

AccountAuthenticator replaces AccountCreationPage as the new base type
for creating an account creation UI.

It would be nice if AccountAuthenticator could be a dialog, but if it
was then AccountsPage would not be able to dynamically load the content
of the next dialog in the Dialog::acceptDestination chain, because
Dialog::acceptDestination cannot be dynamically updated. So instead, we
set the creation dialog's acceptDestination to be an internal Dialog
component that dynamically loads the dialog content, which should be an
AccountAuthenticator-type object.