[jollacomponents-internal] Fix dialog flow for authentication
Now that oauth authentication is done in an embedded view we don't
have to mess around with disabling backNavigation for the oauth
loading screen. Also, we can push the Settings page through a
Replace rather than Push operation so that navigating back from the
Settings page returns to the main accounts view.