[jollacomponents-internal] Ensure services are enabled by default - JB#5223
The previous commit ensured that the account was disabled with all
services until saved, however there were two bugs in the modifications
to the settings dialog: it passed the wrong parameter name during
construction, and it didn't use the parameter during initialisation of
the checked value for each of the toggles.  This commit fixes these
two issues.

Contributes to JB#5223