[jollacomponents-internal] Fix account creation flow - JB#6030
Previously, the settings page could be destroyed prior to the account
creation page being accepted, resulting in the accounts application
getting "stuck" due to invalid page being pushed to the pagestack.
This could occur due to the fact that the "acceptDestination" object
must exist (and is cached by the PageStack) at _creation time_ of the
dialog, instead of at _acceptance time_.

This commit fixes the issue by ensuring that the settings page itself
is created on demand when the creation dialog itself is accepted.

Note 1) this causes a dramatic performance decrease, which needs to
be fixed in the future (by precreating all settings dialogs as
required, and merely setting the account id on account creation).
Note 2) existing accounts ui extension settings pages now need to
change slightly (using a Connections element to detect accept).

Contributes to JB#6030