[jollacomponents-internal] Improve OAuth account creation performance
This commit changes the OAuthAccountAuthenticator to use a custom
account factory type instead of the nemo-qml-plugins-accounts and
nemo-qml-plugins-signon adaptors to create the account.

In doing so, we remove the "account" property from the base
authenticator type, which means we need to add value decoding support
to the encoded key provider.