[sailfish-components-accounts] Allow multiple accounts to be selected when adding new accounts. JB#6966
When multiple accounts are selected for creation, we display each set
of account creation & settings pages one after the other.

When a provider is selected in the picker dialog, the account creation
page for that provider is loaded. When that page becomes active, its
settings page is loaded and set it as the next page. When the settings
page becomes active, we load the account creation page for the next
provider that was selected in the picker, and set it as the next page,
and so it continues until we have gone through all the selected
providers.

The instantiated account creation and settings pages are deleted once
the whole process is completed.