[sailfish-components-accounts] Simple account creation should show error if creation fails. JB#7012
Instead of showing the settings page immediately, show an
intermediate page and then transition to the settings page if
account creation is successful, or show an error otherwise.

This requires accountcreationmanager.js to create the next
account creation page in the queue as soon as the current settings
page is created, so that the user can navigate directly from the
error page to the next account creation page.