[sailfish-accounts] Emit error type with sign in error signal. Contributes to JB#7479
Previously, only a string message was emitted with the error signal.
This commit adds type information, so that clients can determine
whether it was due to (for example) credentials expiring.  This is
needed so that sync adapters can know to create a notification for
the user that they need to update their credentials via the settings
application.

Contributes to JB#7479