[sailfish-components-accounts] Add --update-sync-services option to sailfish-accounts-tool. Contributes to JB#15786
This will be used to upgrade old accounts that used to only have a
generic sync service. In the new world order, we have one service per
sync data type (e.g. contacts, calendars, etc.) so this option can
be used to copy settings from the generic sync service to the
data-type-specific services, and also generic a sync profile for each
of them.