[sailfish-components-accounts] Report better errors. Contributes to JB#15021
This commit ensures that error codes reported by the account during
a signon process are useful even if the code reported by signond is
useless, by inspecting the error message content.

This allows clients to determine if the credentials have expired or
not, by inspecting the error reported from the account.

Contributes to JB#15021