[sailfish-components-accounts] Allow multiple default template sync profiles per service. Contributes to JB#15786
This commit changes the signature of the function which returns
the default sync template for a given account service, to being
able to return multiple sync templates.  This is required if a
single service has several default sync profile templates associated
with it (for example, microblog services generally have multiple).

Contributes to JB#15786