[sailfish-components-accounts] Correctly update sync service profiles and return -1 status on tool error. Contributes to JB#18826
The generic <provider>-sync.service now has template profiles so we
cannot use this characteristic to test whether it is the generic
sync service. Use the service name to test this instead.

Also change the tool to return -1 on error so that it can be tested for
failure when run from a script, and fix the compile flags for
sailfish-accounts-tool.