[sailfish-components-accounts] Don't overwrite existing profiles when sailfish-accounts-tool is run without privileged permissions. Fixes JB#21547
When executing an --update-sync-services or --create-profiles command,
if the tool does not have privileged permissions to read/write within
the sync profiles directory, it thinks the profile does not exist and
so asks msyncd to write a new profile for that account, which causes
the existing profile to be overwritten. So, check the directory is
actually readable before calling updateProfile() via msyncd.