[sailfish-components-accounts-tools] Timeout stuck backup/restore operation. Contributes to JB#27255
This commit ensures that if a backup or restore operation gets stuck,
we time-out the asynchronous request which causes that state.
This can occur, for example, if the signon process hangs during
a signon request.

It also removes some unnecessary threading and relies on the main
thread's event loop to perform processing.

Contributes to JB#27255