[sailfish-components-accounts] Make unit tests more stable. Contributes to JB#27966
Recent changes to the Accounts&SSO libraries mean that the Manager
instance caches (in memory) account data.  When an account is modified
and sync()ed, the synced() signal can now be emitted BEFORE the
Manager-owned cached copy is updated with the modifications.  This
has the unfortunate side effect that attempting to retrieve an account
instance from the Manager does not guarantee that the settings will
be up to date.

This commit merely inserts delays after problematic sync() calls
to ensure that subsequent cache-reads result in the most up-to-date
settings being retrieved, in the unit test.

Note that this issue should not affect Sailfish software adversely,
as the case where multiple codepaths modify the same account within
the same process are very rare.

Once we have either identified the underlying issue in accounts&sso
or ported away from accounts&sso entirely, we should revert this
commit.  In the meantime, this will allow integration to continue.

Contributes to JB#27966