[sailfish-components-accounts] Cancel pending OAuth requests in cancelSignInOperation(). Contributes to JB#28089
Currently, if an account is deleted while it is signing in with an
OAuth session, the OAuth session will go ahead anyway. Signing out of
the identity and removing it isn't enough, if the auth session is still
in progress; the session has to be cancelled explicitly with
AuthSession::cancel().