[sailfish-components-accounts] Don't emit an extra signInError() signal. Contributes to JB#28089
This fixes the tst_Account::credentialsFunctions() test failure.

When SignOn::AuthSession::cancel() is called, it emits
error(Error::SessionCanceled). Ensure account.cpp doesn't pass this
error on as a signInError(), else the signInError() signal is emitted
twice.

We can't move the signInError() emission out of cancelCredentialsOperation()
as the session may not have been started yet when cancel() is called.