[sailfish-accounts-tool] Ensure we disable accounts when availability is removed. Contributes to JB#28750
This commit ensures that accounts are disabled appropriately when
required.  Previously, an account may not have been disabled if the
KeyProviderAvailable key was not set on the account - which would
result in returning false for that key (null evaluates to false).
In that case, this "spurious" false value would match the wanted
enabled state (ie, false) and thus the tool had skipped the disable
operation.

With this commit, it only skips the disable operation if the enabled
status is also correct.

Contributes to JB#28750