[sailfish-components-accounts] Ensure A&SSO emits change signals on enable/disable. Contributes to JB#31375
The Manager class from Accounts&SSO will only emit change signals like
accountUpdated() if the Manager is constructed with a service type
filter.

This means that clients who listen for change signals will not be
notified that the entire account has been enabled or disabled, if
only the global enabled value is updated in the account.

This commit ensures that we also modify a dummy trigger value in every
service of the account when setting the account's enablement status
via the AccountModel, to ensure that other clients will be notified
of the change.

Contributes to JB#31375