[sailfish-components-accounts] Improve backup/restore handling.  Contributes to JB#33183
This commit ensures that we still generate sync profiles when
attempting to restore an account which does not have a "-sync"
service.  Some accounts (like Dropbox) don't have a -sync service,
but still require sync profiles (for Images etc) to be generated.

It also ensures that when we are backing up accounts, we back up
credentials associated with the account even if it does not have a
-sync or -signon service, for the same reason.

Finally, it ensures that we don't report an error if the vk.Sharing
sync profile cannot be generated, because that profile may not exist.

Contributes to JB#33183 and JB#33240