[sailfish-components-accounts-tool] Don't fail restore if sync template missing. Contributes to JB#36350
Previously, if a sync profile for the expected sync template could not
be created, the restore process would fail.  However, this doesn't
take into account the case where a backup is made on a device with
some sync template, and restored to a device without that template.
One common case is if the jolla-email app is uninstalled from the
device.

This commit ensures that in that case, the restore continues and
does not treat the failure to create the sync template as an
unrecoverable error.

Contributes to JB#36350