[sailfish-components-accounts] Don't migrate "new" cdav accounts. Contributes to MER#1822
Previously, we always did a backup+restore cycle of cdav accounts,
even if no migration of account settings was required.
This results in loss of continuity of sync history (since the account
id will have changed due to the backup/restore cycle), which can
cause the database to be clean-synced unnecessarily.

This commit ensures that we only do a backup/restore cycle for those
accounts which still use the "old" service names, thus requiring
migration to the new services.

Contributes to MER#1822