[components-accounts] Update only account values that have changed. Contributes to JB#44594
When configuration values are set on an account or service, they will
only be committed on a sync if the values have actually changed since
the previous sync. This ensures that any changes made by other processes
in the meantime are not overwritten.

A new AccountValue QObject is also provided, which allows account values
to be presented live (i.e. updated dynamically).

If set on another property, the property will be updated as the account
setting changes. If the property is changed (e.g. by the user) then the
Account value will be set and the live link disconnected (i.e. the value
will no longer be updated live).