[components-accounts] Always update service enabled configuration value. Contributes to JB#51570
When the user created a new Google account the service enabled state
chosen on the settings page wasn't sticking (all services were enabled
independent of the user's choice).

This was because the service enabled state didn't update directly during
initialisation, instead being stored in the "enabled" configuration
value. Since the configuration value was only being updated if the
actual value needed changing, and the actual value was always "false",
the configuration value would never be set to "false".

This change ensures the configuration value is always updated
independent of the actual value, ensuring that the user's choice sticks
when the account is created.