[sailfish-components-accounts] Add fallback url for authentication check. Contributes to JB#52704
Currently, if the authenticated request fails, we immediately assume
that the credentials were incorrect.  However, it can also fail if
the server doesn't support authenticated requests to the root path.

This commit ensures that if the request to the root path fails, a
second authenticated request is sent to the .well-known endpoint.
If that one also fails, then we assume that the credentials were
incorrect.