[components-accounts] Set CredentialsNeedUpdate flag on account after restoration. JB#48032
After a backup is restored, cloud accounts credentials are lost so
are no longer able to fetch data until the user signs in again for
each account. Normally this loss of credentials is not detected until
an account sign-in is attempted again (e.g. when a sync is done), at
which point the sign-in fails and the CredentialsNeedUpdate flag is
set to notify the UI that the user needs to sign in.

To make this more clear to the user, set the CredentialsNeedUpdate
as soon as the account is restored.