[tests] Ensure properties change before testing in unit tests. Fixes JB#57256
Account syncs cause properties to be written to the accounts database
and propagated to other instances of the Account accessor. However the
propagation is performed asynchronously using dbus signals, so
completion of the sync doesn't guarantee that the changes have fully
propagated.

The unit tests were failing because in some cases the checks were
waiting on an account synced signal, by which point the signals had been
sent, but not necessarily received. The property checks on other account
instances were therefore failing.

This change adds in some additional signal spies to the unit tests to
ensure the property changes have fully propaged before checking.